offline user

+++ cim-common/src/main/java/com/crossoverjie/cim/common/req/BaseRequest.java
@@ -10,14 +10,11 @@ import io.swagger.annotations.ApiModelProperty;
  */
 public class BaseRequest {
 
-
     @ApiModelProperty(required=false, value="唯一请求号", example = "1234567890")
     private String reqNo;
 
     @ApiModelProperty(required=false, value="当前请求的时间戳", example = "0")
-    private int timeStamp;
-
-
+    private long timeStamp;
 
     public BaseRequest() {
         this.setTimeStamp((int)(System.currentTimeMillis() / 1000));
@@ -31,11 +28,11 @@ public class BaseRequest {
         this.reqNo = reqNo;
     }
 
-    public int getTimeStamp() {
+    public long getTimeStamp() {
         return timeStamp;
     }
 
-    public void setTimeStamp(int timeStamp) {
+    public void setTimeStamp(long timeStamp) {
         this.timeStamp = timeStamp;
     }
